Neatens eslint imports, moves node dependent code into node directory within src
