+++ package.json
@@ -30,6 +30,7 @@
     "eslint-config-prettier": "^10.1.8",
     "eslint-import-resolver-typescript": "^4.4.4",
     "eslint-plugin-import": "^2.32.0",
+    "globals": "^16.3.0",
     "netlify-cli": "^23.1.3",
     "postcss": "^8.5.6",
     "prettier": "^3.6.2",
